made OSCInputStream and OSCOutputStream public
