+++ modules/juce_osc/juce_osc.cpp
@@ -43,3 +43,5 @@
 #include "osc/juce_OSCBundle.cpp"
 #include "osc/juce_OSCReceiver.cpp"
 #include "osc/juce_OSCSender.cpp"
+#include "osc/juce_OSCInputStream.cpp"
+#include "osc/juce_OSCOutputStream.cpp"
